backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-multipart
python-dotenv
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
requests